project1/requirement.txt
bcrypt==3.1.4
cffi==1.11.2
Django==1.8
djangorestframework==3.6.2
Faker==0.8.7
ipaddress==1.0.19
pycparser==2.18
python-dateutil==2.6.1
six==1.11.0
text-unidecode==1.1
